moving temporary components to a temporary folder to make it more obvious they won't always be used
