+++ services/admin/src/components/temporary/ManifestTable.svelte
